==> Automated/Code/extremes_code/step1.dqcheck.ipynb <==
import pandas as pd
import numpy as np
import csv
import calendar
import plotly.graph_objs as go
import plotly.express as px
from plotly.subplots import make_subplots

import dash
import dash_table
import dash_core_components as dcc
import dash_html_components as html
import datetime as dt

import os,sys,inspect
from dash.dependencies import Input, Output, State

current_dir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
parent_dir = os.path.dirname(current_dir)
sys.path.insert(0, parent_dir) 
from environmentVariables import environment_variables

user_key=environment_variables['eshta']
pd.set_option('display.precision', 12)

data=pd.read_csv(user_key['extremes_raw_unclean']+'weatherstats_toronto_normals_daily-1990-2020.csv')
data2= pd.read_csv(user_key['extremes_raw_unclean']+'weatherstats-toronto-daily-1990-2020.csv')
data3= pd.read_csv(user_key['extremes_raw_unclean']+'weatherstats_toronto_normals_monthly-1990-2020.csv')


#clean up the data here

#1. Check for nulls or zero values
#2. Check if all days are there

nullCounter = 0
zeroCounter = 0
blankCounter = 0
counter = 0
row_iterator = data.iterrows()


listofDates= []
listofColumns= []
listofMonths= []
listofYears= []


for i, row in row_iterator:
        if counter != len(data):
            for column in range(0,53):
                #print((data.at[i,data.columns[column]]))                                                                                    
                if (pd.isnull(data.at[i,data.columns[column]])):
                    #create a list of dates
                    #print('test')
                    #only append if column has _v in it
                    if ( '_v'in data.columns[column] ):
                        
                        listofDates.append(data.at[i,'date'])
                        listofMonths.append(pd.to_datetime(data.at[i,'date']).month)
                        listofYears.append(pd.to_datetime(data.at[i,'date']).year)
                        listofColumns.append(data.columns[column])
                        nullCounter += 1
                    
                    
                    
                
                    
        counter += 1   
        
        
#print(listofColumns)                        
Daily_Report_normals = pd.DataFrame()


Daily_Report_normals.insert(0, 'Date', listofDates)
Daily_Report_normals.insert(1, 'Month', listofMonths)
Daily_Report_normals.insert(2, 'Year', listofYears)
Daily_Report_normals.insert(3, 'Column', listofColumns)



Daily_Report_normals = (Daily_Report_normals.groupby(['Column', 'Year','Month']).size().reset_index(name='counts'))



nullCounter = 0
zeroCounter = 0
blankCounter = 0
counter = 0
row_iterator = data2.iterrows()


listofDates= []
listofColumns= []
listofMonths= []
listofYears= []


for i, row in row_iterator:
        if counter != len(data2):
            for column in range(0,70):
                if (pd.isnull(data2.at[i,data2.columns[column]])):
  
                    listofDates.append(data2.at[i,'date'])
                    listofMonths.append(pd.to_datetime(data2.at[i,'date']).month)
                    listofYears.append(pd.to_datetime(data2.at[i,'date']).year)
                    listofColumns.append(data2.columns[column])
                    nullCounter += 1
                    
                    
                    
                
                    
        counter += 1   
        
        
#print(listofColumns)                        
Daily_Report_Observed = pd.DataFrame()


Daily_Report_Observed.insert(0, 'Date', listofDates)
Daily_Report_Observed.insert(1, 'Month', listofMonths)
Daily_Report_Observed.insert(2, 'Year', listofYears)
Daily_Report_Observed.insert(3, 'Column', listofColumns)


Daily_Report_Observed = (Daily_Report_Observed.groupby(['Column', 'Year','Month']).size().reset_index(name='counts'))



nullCounter = 0
zeroCounter = 0
blankCounter = 0
counter = 0
row_iterator = data3.iterrows()


listofDates= []
listofColumns= []
listofMonths= []
listofYears= []


for i, row in row_iterator:
        if counter != len(data3):
            for column in range(0,53):
                if (pd.isnull(data3.at[i,data3.columns[column]])):
                    #only append if column has _v in it
                    if ( '_v'in data3.columns[column] ):
                        
                        listofDates.append(data3.at[i,'date'])
                        listofMonths.append(pd.to_datetime(data3.at[i,'date']).month)
                        listofYears.append(pd.to_datetime(data3.at[i,'date']).year)
                        listofColumns.append(data3.columns[column])
                        nullCounter += 1
                    
                    
                    
                
                    
        counter += 1   
        
        
#print(listofColumns)                        
Monthly_Report_Normals = pd.DataFrame()


Monthly_Report_Normals.insert(0, 'Date', listofDates)
Monthly_Report_Normals.insert(1, 'Month', listofMonths)
Monthly_Report_Normals.insert(2, 'Year', listofYears)
Monthly_Report_Normals.insert(3, 'Column', listofColumns)


Monthly_Report_Normals = (Monthly_Report_Normals.groupby(['Column', 'Year','Month']).size().reset_index(name='counts'))


Daily_Report_normals.to_csv(user_key['extremes_raw_dq']+'daily_report_normals.csv', index=None)
Daily_Report_Observed.to_csv(user_key['extremes_raw_dq']+'daily_report_observed.csv', index=None)
Monthly_Report_Normals.to_csv(user_key['extremes_raw_dq']+'monthly_report_normals.csv', index=None)
